pyproject.toml
[tool.poetry]
name = "ez-housing"
version = "0.1.0"
description = ""
authors = ["Giuseppe <giuseppe.fur96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.1"
selenium = "^4.19.0"
bs4 = "^0.0.2"
pymongo = "^4.6.3"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
apache-airflow = "^2.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
